Added colors in the Video block
