+++ src/gutenberg/blocks/video/styles/style.scss
@@ -186,6 +186,10 @@
         width: 100%;
         height: 100%;
       }
+
+      video {
+        background: var(--gkt-fullscreen-video--video__background);
+      }
     }
 
     &-close {
